# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TogetherPlan API",
        "description": "API documentation for the TogetherPlan project. This API allows users to manage events, participants, voting on dates, groups, and notifications.",
        "contact": {
            "name": "TogetherPlan Support",
            "email": "support@togetherplan.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.togetherplan.com",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/votes": {
            "post": {
                "tags": [
                    "Votes"
                ],
                "summary": "Submit a vote for a date option",
                "description": "Allows an authenticated user to vote 'yes', 'maybe', or 'no' on a date option. Recalculates the best date for the event.",
                "operationId": "storeVote",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date_option_id",
                                    "vote"
                                ],
                                "properties": {
                                    "date_option_id": {
                                        "description": "ID of the date option to vote on",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "vote": {
                                        "description": "Type of vote",
                                        "type": "string",
                                        "enum": [
                                            "yes",
                                            "maybe",
                                            "no"
                                        ],
                                        "example": "yes"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vote submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vote submitted successfully!"
                                        },
                                        "vote": {
                                            "$ref": "#/components/schemas/Vote"
                                        },
                                        "best_date_id": {
                                            "description": "ID of the current best date for the event",
                                            "type": "integer",
                                            "example": 16
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User has already voted for this date option",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have already voted for this date."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The date option id field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vote": {
                "title": "Vote",
                "description": "Represents a vote cast by a user for a date option in an event",
                "properties": {
                    "id": {
                        "description": "Vote ID",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the user who voted",
                        "type": "integer",
                        "example": 11
                    },
                    "event_id": {
                        "description": "ID of the event associated with the vote",
                        "type": "integer",
                        "example": 6
                    },
                    "date_option_id": {
                        "description": "ID of the date option voted for",
                        "type": "integer",
                        "example": 17
                    },
                    "vote": {
                        "description": "Vote choice",
                        "type": "string",
                        "enum": [
                            "yes",
                            "maybe",
                            "no"
                        ],
                        "example": "yes"
                    },
                    "points": {
                        "description": "Numeric points associated with the vote",
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Laravel Sanctum API token for authentication",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}